feat(ansible): create structure for ansible files
Creation of basic ansible files and structure for folder. Trying to
keep smaller playbooks to quickly update and debug if needed.

If there is any new role, just create a folder, and within it, a new
folder named task, and inside there, create the main.yml for that
setting.

There are other possible folder, handlers and vars/defaults, here is the
explanation:

-tasks: for tasks(updating, installing, downloading, etc)
-handlers: if there is any service that needs to be restarted(i.e.:
nginx)
-vars/defualts: for managing versions and default settings.